+++ src/common/interfaces.ts
@@ -1,5 +1,7 @@
-export interface Session {
+interface Session {
     category: string;
     subject: string;
     time: number;
 }
+
+export type { Session }
