+++ src/index.ts
@@ -1,8 +1,8 @@
 // Imports use relative file paths or Node.js package names
 import './styles/styles.css';
-import {initNav} from "./utility.ts";
-import { startTimer, pauseTimer, stopTimer} from "./timer.ts";
-import { startBtnEl, pauseBtnEl, stopBtnEl} from "./dom-utils.ts";
+import { initNav } from "./common/utility.ts";
+import { startTimer, pauseTimer, stopTimer } from "./timer/timer.ts";
+import { startBtnEl, pauseBtnEl, stopBtnEl } from "./common/dom-utils.ts";
 
 initNav();
